exclude symbols of all static libraries CNTKLIBRARY_LIB is using (e.g. protobuf)
this is required as CNTK is using protobuf 3.1 static library that asserts when
the system wide installed of protobuf is at higher / incompatible version.
Specifically, on Linux the method google::protobuf::internal::VerifyVersion that is invoked in *.pb.cc (protobuf generated files) and
exported in libCntk.Core-2.6.so, will be relocated by the linker to the installed version used by the main application – hence may fail validation if they are incompatible.